Int width / height changed in float

+++ Code/glimac/include/project/Ghost.hpp
@@ -10,7 +10,7 @@ private:
 public: 
 
     // constructor
-    Ghost(int, int, int, int, int, int, Orientation);
+    Ghost(int, int, float, float, int, int, Orientation);
     Ghost();
 
     // getter
